Improve character system decomp matches
Add and refine implementations across CharBone, CharBlendBone,
CharEyes, CharHair, CharUpperTwist, CharGuitarString, CharDriver,
CharClipGroup, CharServoBone, CharForeTwist, Waypoint, and others.
Fixes sign/unsigned comparisons, variable ordering, and missing
function stubs.

+++ src/system/char/CharEyes.h
@@ -153,7 +153,7 @@ protected:
     RndOverlay *mEyeStatusOverlay; // 0xd0
     int mInterestFilterFlags; // 0xd4
     Vector3 unkd8; // 0xd8
-    int unke8;
+    float unke8;
     float unkec;
     float unkf0;
     int unkf4;
